use lemon modal, remove projects only if no other licenses available

+++ ee/api/license.py
@@ -2,6 +2,7 @@ import requests
 from django.conf import settings
 from django.db.models import QuerySet
 from django.shortcuts import get_object_or_404
+from django.utils.timezone import now
 from rest_framework import mixins, request, serializers, viewsets
 from rest_framework.response import Response
 
@@ -51,15 +52,17 @@ class LicenseViewSet(
         validation = requests.post("https://license.posthog.com/licenses/deactivate", data={"key": license.key})
         validation.raise_for_status()
 
-        teams = Team.objects.exclude(is_demo=True).order_by("pk")[1:]
-        for team in teams:
-            team.delete()
+        has_another_valid_license = License.objects.filter(valid_until__gte=now()).exclude(pk=pk).exists()
+        if not has_another_valid_license:
+            teams = Team.objects.exclude(is_demo=True).order_by("pk")[1:]
+            for team in teams:
+                team.delete()
 
-        #  delete any organization where we've deleted all teams
-        # there is no way in the interface to create multiple organizations so we won't bother informing people that this is happening
-        for organization in Organization.objects.all():
-            if organization.teams.count() == 0:
-                organization.delete()
+            #  delete any organization where we've deleted all teams
+            # there is no way in the interface to create multiple organizations so we won't bother informing people that this is happening
+            for organization in Organization.objects.all():
+                if organization.teams.count() == 0:
+                    organization.delete()
 
         license.delete()
